MatrixPilot_trunk: purely cosmetic changes to simplify difference analysis between trunk and the UDB5 and AUAV3 branches. Introduction of several static labels and a few unused defines however no functional changes.
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@2469 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/flightplan-logo.c
@@ -26,11 +26,11 @@
 
 
 struct logoInstructionDef {
-	unsigned int cmd		:  6 ;
-	unsigned int do_fly		:  1 ;
-	unsigned int use_param	:  1 ;
-	unsigned int subcmd		:  8 ;
-	int arg		: 16 ;
+	uint16_t cmd		:  6 ;
+	uint16_t do_fly		:  1 ;
+	uint16_t use_param	:  1 ;
+	uint16_t subcmd		:  8 ;
+	int16_t arg			: 16 ;
 } ;
 
 #define PLANE				0
